requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.0
facebook-sdk==3.1.0
fastapi==0.91.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pydantic==1.10.4
pytest==7.2.1
python-dotenv==0.21.1
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
SQLAlchemy==2.0.3
starlette==0.24.0
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
